Move govuk_content_schemas_path to a shared module
I want to reuse this code in another class, and it isn't schema/validation
specific anyway.
