+++ lib/govuk-content-schema-test-helpers.rb
@@ -1,4 +1,5 @@
 require 'govuk-content-schema-test-helpers/configuration'
+require 'govuk-content-schema-test-helpers/util'
 require 'govuk-content-schema-test-helpers/validator'
 
 module GovukContentSchemaTestHelpers
