+++ lib/govuk-content-schema-test-helpers/util.rb
@@ -1,0 +1,7 @@
+module GovukContentSchemaTestHelpers
+  module Util
+    def self.govuk_content_schemas_path
+      ENV['GOVUK_CONTENT_SCHEMAS_PATH'] || '../govuk-content-schemas'
+    end
+  end
+end
