+++ lib/govuk-content-schema-test-helpers/validator.rb
@@ -4,14 +4,10 @@ module GovukContentSchemaTestHelpers
   class ImproperlyConfiguredError < RuntimeError; end
 
   class Validator
-    def self.govuk_content_schemas_path
-      ENV['GOVUK_CONTENT_SCHEMAS_PATH'] || '../govuk-content-schemas'
-    end
-
     def self.schema_path(schema_name)
       project_root = GovukContentSchemaTestHelpers.configuration.project_root
       schema_type = GovukContentSchemaTestHelpers.configuration.schema_type
-      File.join(project_root, self.govuk_content_schemas_path, "/formats/#{schema_name}/#{schema_type}/schema.json").to_s
+      File.join(project_root, Util.govuk_content_schemas_path, "/formats/#{schema_name}/#{schema_type}/schema.json").to_s
     end
 
     # schema_name should be a string, such as 'finder'
@@ -19,7 +15,7 @@ module GovukContentSchemaTestHelpers
     def initialize(schema_name, document)
       @schema_path = GovukContentSchemaTestHelpers::Validator.schema_path(schema_name)
 
-      govuk_content_schemas_path = GovukContentSchemaTestHelpers::Validator.govuk_content_schemas_path
+      govuk_content_schemas_path = Util.govuk_content_schemas_path
       if !Pathname(govuk_content_schemas_path).exist?
         message = "Dependency govuk-content-schemas cannot be found at: #{govuk_content_schemas_path}."
         message += " Clone it to that directory, or set GOVUK_CONTENT_SCHEMAS_PATH (see README.md for details)."
